move modules_in_pkg and load_method into cli.py as it's the only callsite
