+++ tests/cli/test_utils.py
@@ -239,15 +239,6 @@ def test_lazy_choices_completer():
     assert completer(prefix="") == ["1", "2", "3"]
 
 
-def test_modules_in_pkg():
-    from paasta_tools.cli import cmds
-
-    ret = tuple(utils.modules_in_pkg(cmds))
-    assert "__init__" not in ret
-    assert "cook_image" in ret
-    assert "list_clusters" in ret
-
-
 @mock.patch("paasta_tools.cli.utils.INSTANCE_TYPE_HANDLERS", dict(), autospec=None)
 @mock.patch("paasta_tools.cli.utils.validate_service_instance", autospec=True)
 def test_get_instance_config_by_instance_type(mock_validate_service_instance,):
